refactor: rename and replace cart page components
