+++ src/components/CartList/QuantityControls/index.ts
@@ -1,0 +1,2 @@
+import QuantityControls from './QuantityControls.tsx';
+export { QuantityControls };
