+++ src/components/CartList/QuantityControls/quantityControls.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common/';
+@import 'common/index';
 
 .quantity__controls {
   height: 4.4rem;
