+++ src/components/CartList/index.ts
@@ -1,0 +1,2 @@
+import CartList from './CartList.tsx';
+export { CartList };
