+++ src/components/CartTotal/PromoCodes/index.ts
@@ -1,0 +1,2 @@
+import PromoCodes from './PromoCodes.tsx';
+export { PromoCodes };
