+++ src/components/CartTotal/index.ts
@@ -1,0 +1,2 @@
+import CartTotal from './CartTotal.tsx';
+export { CartTotal };
