+++ src/components/ProductInfo/ProductDescription/index.ts
@@ -1,0 +1,2 @@
+import ProductDescription from './ProductDescription.tsx';
+export { ProductDescription };
