+++ src/components/ProductInfo/ProductDescription/productDescription.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import '../../../styles/common/index';
 
 .info {
   width: 100%;
