+++ src/components/ProductInfo/ProductTitle/index.ts
@@ -1,0 +1,2 @@
+import ProductTitle from './ProductTitle.tsx';
+export { ProductTitle };
