+++ src/components/ProductInfo/ProductTitle/productTitle.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import '../../../styles/common/index';
 
 .artist {
   font-size: 4rem;
