+++ src/components/ProductInfo/index.ts
@@ -1,0 +1,2 @@
+import ProductInfo from './ProductInfo.tsx';
+export { ProductInfo };
