+++ src/components/ProductInfo/productInfo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import '../../styles/common/index';
 
 .product__details {
   max-width: 50vw;
