--- src/pages/Cart/CartList/QuantityControls/index.ts
@@ -1,2 +1,0 @@
-import QuantityControls from './QuantityControls';
-export { QuantityControls };
