--- src/pages/Cart/CartList/index.ts
@@ -1,2 +1,0 @@
-import CartList from './CartList';
-export { CartList };
