--- src/pages/Cart/PromoCodes/index.ts
@@ -1,2 +1,0 @@
-import PromoCodes from './PromoCodes';
-export { PromoCodes };
