--- src/pages/Cart/Total/index.ts
@@ -1,2 +1,0 @@
-import Total from './Total';
-export { Total };
