--- src/pages/Product/ProductDescription/index.ts
@@ -1,2 +1,0 @@
-import ProductDescription from './ProductDescription';
-export { ProductDescription };
