--- src/pages/Product/ProductInfo/index.ts
@@ -1,2 +1,0 @@
-import ProductInfo from './ProductInfo';
-export { ProductInfo };
