--- src/pages/Product/ProductTitle/index.ts
@@ -1,2 +1,0 @@
-import ProductTitle from './ProductTitle';
-export { ProductTitle };
